fix: コンポーネントの整理 Fix #208
