+++ src/components/index.ts
@@ -4,9 +4,9 @@
 
 export * from './atoms/SvgIcon';
 export * from './cards/ExpandCardContent';
-export * from './layouts/CookieAlert';
 export * from './layouts/Footer';
 export * from './layouts/Header';
+export * from './organisms/CookieAlertContent';
 export * from './sections/Contacts';
 export * from './sections/Hello';
 export * from './sections/Histories';
@@ -15,6 +15,7 @@ export * from './sections/Projects';
 export * from './sections/Qualifications';
 export * from './sections/Skills';
 export * from './sections/WhatICanDos';
+export * from './templates/CookieAlert';
 export * from './templates/Head';
 export * from './templates/Layout';
 export * from './utilities/InlineBlock';
