+++ src/components/templates/Head.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Helmet } from 'gatsby-plugin-react-i18next';
-import { useBuildTime, useSiteMetadata } from '../../hooks';
+import { useBuildTime, useSiteMetadata } from 'src/hooks';
 
 type Props = {
   title?: string;
